Added tabview for search option. Now user can select to do content search or user search by the tabview on the top of search view
