+++ android/app/src/main/java/com/example/learningplatform/ContentSearchFragment.java
@@ -1,0 +1,13 @@
+package com.example.learningplatform;
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import androidx.fragment.app.Fragment;
+
+public class ContentSearchFragment extends Fragment {
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.activity_search, container, false);
+    }
+}
