+++ android/app/src/main/java/com/example/learningplatform/SearchTabAdapter.java
@@ -1,0 +1,36 @@
+package com.example.learningplatform;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
+import java.util.ArrayList;
+import java.util.List;
+
+public class SearchTabAdapter extends FragmentPagerAdapter {
+    private final List<Fragment> mFragmentList = new ArrayList<>();
+    private final List<String> mFragmentTitleList = new ArrayList<>();
+
+    public SearchTabAdapter(FragmentManager fragmentManager){
+        super(fragmentManager);
+    }
+    @Override
+    public Fragment getItem(int position) {
+        return mFragmentList.get(position);
+    }
+    public void addFragment(Fragment fragment, String title) {
+        mFragmentList.add(fragment);
+        mFragmentTitleList.add(title);
+    }
+
+    @Nullable
+    @Override
+    public CharSequence getPageTitle(int position) {
+        return mFragmentTitleList.get(position);
+    }
+    @Override
+    public int getCount() {
+        return mFragmentList.size();
+    }
+}
+
+
